setup routing, move queries in ConfigurationIndex to parent resource route
