+++ ui/lib/pki/addon/components/page/pki-configuration-edit.hbs
@@ -1,0 +1,3 @@
+<div class="box is-sideless is-fullwidth is-marginless">
+  Edit stuff happens here
+</div>
