+++ ui/lib/pki/addon/components/page/pki-configuration-edit.ts
@@ -1,0 +1,8 @@
+/**
+ * Copyright (c) HashiCorp, Inc.
+ * SPDX-License-Identifier: MPL-2.0
+ */
+
+import Component from '@glimmer/component';
+
+export default class PkiConfigurationEditComponent extends Component {}
