+++ ui/lib/pki/addon/decorators/check-config.js
@@ -4,6 +4,7 @@
  */
 
 import Route from '@ember/routing/route';
+import { inject as service } from '@ember/service';
 
 /**
  * the overview, roles, issuers, certificates, and key routes all need to be aware of the whether there is a config for the engine
@@ -21,6 +22,7 @@ export function withConfig() {
       return SuperClass;
     }
     return class CheckConfig extends SuperClass {
+      @service secretMountPath;
       shouldPromptConfig = false;
 
       async beforeModel() {
