+++ ui/lib/pki/addon/templates/configuration/edit.hbs
@@ -1,1 +1,13 @@
-configuration.edit
+<PageHeader as |p|>
+  <p.top>
+    <Page::Breadcrumbs @breadcrumbs={{this.breadcrumbs}} />
+  </p.top>
+  <p.levelLeft>
+    <h1 class="title is-3">
+      <Icon @name="file-text" @size="24" class="has-text-grey-light" />
+      Edit PKI configuration
+    </h1>
+  </p.levelLeft>
+</PageHeader>
+
+<Page::PkiConfigurationEdit />
